refactor logging / util packages to env
